refactor: 검색 결과 아이템 컴포넌트 통일
